+++ src/models/ContentFolder.ts
@@ -5,4 +5,5 @@ export interface ContentFolder {
   excludeSubdir?: boolean;
   previewPath?: string;
   filePrefix?: string;
+  contentTypes?: string[];
 }